requirements.txt
numpy
pandas
xarray
eccodes
torch --index-url https://download.pytorch.org/whl/cpu
scikit-learn
boto3
watchtower
ecmwf-opendata